streamlit_app/requirements.txt
streamlit
pandas
plotly
google-cloud-bigquery
